web/requirements.txt
setuptools<71
flask==3.0.3
gunicorn==22.0.0
requests==2.32.3
google-auth==2.29.0
google-auth-httplib2==0.2.0

# OpenTelemetry — exports via HTTP to the local managed OTel Collector DaemonSet
opentelemetry-sdk==1.24.0
opentelemetry-exporter-otlp-proto-http==1.24.0
opentelemetry-resourcedetector-gcp==1.5.0a0
opentelemetry-propagator-gcp==1.5.0a0
opentelemetry-instrumentation-flask==0.45b0
opentelemetry-instrumentation-requests==0.45b0
